Reset streams with slow readers.
Currently, slow readers can grind the entire connection to a halt by never
reading from a stream and allowing it's read buffer to fill up.

This commit resets streams in this condition if writing to the read buffer takes
more than 15 seconds (if something fails to read on a stream for over 15
seconds, we have a problem...).